requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==4.4.2
filelock==3.16.0
fsspec==2024.9.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
pillow==10.4.0
proglog==0.1.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
sympy==1.13.2
tiktoken==0.7.0
torch==2.2.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
